refactor(tangle-dapp): Move blueprints components to shared ui
