+++ apps/tangle-dapp/app/blueprints/BlueprintListing.tsx
@@ -10,6 +10,7 @@ import {
   TableOptions,
   useReactTable,
 } from '@tanstack/react-table';
+import BoostedChip from '@webb-tools/tangle-shared-ui/components/blueprints/BoostedChip';
 import { SkeletonLoader } from '@webb-tools/webb-ui-components';
 import Button from '@webb-tools/webb-ui-components/components/buttons/Button';
 import { fuzzyFilter } from '@webb-tools/webb-ui-components/components/Filter/utils';
@@ -30,7 +31,6 @@ import useBlueprintListing from '../../data/blueprints/useBlueprintListing';
 import { PagePath } from '../../types';
 import type { Blueprint } from '../../types/blueprint';
 import isSubstrateAddress from '../../utils/isSubstrateAddress';
-import BoostedChip from './BoostedChip';
 
 const columnHelper = createColumnHelper<Blueprint>();
