+++ apps/tangle-dapp/app/blueprints/[id]/BlueprintHeader.tsx
@@ -1,4 +1,5 @@
 import { ArrowRight, GithubFill } from '@webb-tools/icons';
+import BoostedChip from '@webb-tools/tangle-shared-ui/components/blueprints/BoostedChip';
 import Button from '@webb-tools/webb-ui-components/components/buttons/Button';
 import { SocialChip } from '@webb-tools/webb-ui-components/components/SocialChip';
 import { Typography } from '@webb-tools/webb-ui-components/typography/Typography';
@@ -8,7 +9,6 @@ import { twMerge } from 'tailwind-merge';
 
 import { EMPTY_VALUE_PLACEHOLDER } from '../../../constants';
 import type { Blueprint } from '../../../types/blueprint';
-import BoostedChip from '../BoostedChip';
 
 interface BlueprintHeaderProps {
   blueprint: Blueprint;
