+++ libs/tangle-shared-ui/src/assets/blueprints/top-banner.png
